+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Version 0.4.0
 
+### Decembre 1, 2020
+- Add color maps and palettes (dynamic and static), with examples, tests, and predefined cmaps.
+- Plot now supports palette as objects or iterable. Defaults to the 'holi' palette.
+
 ### November 19, 2020
 - Color now supports all models from the xcolor package (i.e. rgb, HTML, hsb, etc.)
 - Factored the Axis environement into a standalone class in prevision of adding subplots.